[ehealth] fix organization and problems in comments
