+++ examples/biomedical/ernie_health/run_pretrain.py
@@ -13,9 +13,6 @@
 # limitations under the License.
 
 import argparse
-import collections
-import itertools
-import logging
 import os
 import io
 import random
@@ -23,24 +20,18 @@ import time
 import json
 import copy
 from collections import defaultdict
-from functools import partial
-from concurrent.futures import ThreadPoolExecutor
 
 import numpy as np
-from tensorboardX import SummaryWriter
-
 import paddle
 import paddle.distributed as dist
-from paddle.io import DataLoader, Dataset
-
-from paddlenlp.transformers import ErnieHealthForTotalPretraining, ElectraModel, ErnieHealthPretrainingCriterion
+from paddlenlp.transformers import ErnieHealthForTotalPretraining, ElectraModel
 from paddlenlp.transformers import ErnieHealthDiscriminator, ElectraGenerator
-from paddlenlp.transformers import ElectraTokenizer
+from paddlenlp.transformers import ElectraTokenizer, ErnieHealthPretrainingCriterion
 from paddlenlp.transformers import LinearDecayWithWarmup
+from paddlenlp.utils.log import logger
+from visualdl import LogWriter
 
-FORMAT = '%(asctime)s-%(levelname)s: %(message)s'
-logging.basicConfig(level=logging.INFO, format=FORMAT)
-logger = logging.getLogger(__name__)
+from dataset import MedicalCorpus, DataCollatorForErnieHealth, create_dataloader
 
 MODEL_CLASSES = {
     "ernie-health": (ErnieHealthForTotalPretraining, ElectraTokenizer),
@@ -50,12 +41,6 @@ MODEL_CLASSES = {
 def parse_args():
     parser = argparse.ArgumentParser()
 
-    parser.add_argument(
-        "--model_type",
-        default="ernie-health",
-        type=str,
-        help="Model type selected in the list: " +
-        ", ".join(MODEL_CLASSES.keys()))
     parser.add_argument(
         "--model_name_or_path",
         default="ernie-health-chinese",
@@ -83,14 +68,14 @@ def parse_args():
         "--max_seq_length",
         default=512,
         type=int,
-        help="max length of each sequence")
+        help="The max length of each sequence")
     parser.add_argument(
-        "--train_batch_size",
-        default=256,
-        type=int,
-        help="Batch size per GPU/CPU for training.", )
+        "--mlm_prob",
+        default=0.15,
+        type=float,
+        help="The probability of tokens to be sampled as masks.")
     parser.add_argument(
-        "--eval_batch_size",
+        "--batch_size",
         default=256,
         type=int,
         help="Batch size per GPU/CPU for training.", )
@@ -110,7 +95,7 @@ def parse_args():
         type=float,
         help="Epsilon for Adam optimizer.")
     parser.add_argument(
-        "--num_train_epochs",
+        "--num_epochs",
         default=100,
         type=int,
         help="Total number of training epochs to perform.", )
@@ -118,7 +103,7 @@ def parse_args():
         "--max_steps",
         default=-1,
         type=int,
-        help="If > 0: set total number of training steps to perform. Override num_train_epochs.",
+        help="If > 0: set total number of training steps to perform. Override num_epochs.",
     )
     parser.add_argument(
         "--warmup_steps",
@@ -153,7 +138,7 @@ def parse_args():
         choices=["cpu", "gpu"],
         help="The device to select to train the model, is must be cpu/gpu.")
     parser.add_argument(
-        "--seed", type=int, default=42, help="random seed for initialization")
+        "--seed", type=int, default=1000, help="random seed for initialization")
     args = parser.parse_args()
     return args
 
@@ -177,214 +162,6 @@ class WorkerInitObj(object):
         random.seed(self.seed + id)
 
 
-class MedicalCorpus(paddle.io.Dataset):
-    def __init__(self, data_path, tokenizer):
-        self.data_path = data_path
-        self.tokenizer = tokenizer
-        # Add ids for suffixal chinese tokens in tokenized text, e.g. '##度' in '百度'.
-        # It should coincide with the vocab dictionary in preprocess.py.
-        suffix_vocab = {}
-        for idx, token in enumerate(range(0x4E00, 0x9FA6)):
-            suffix_vocab[len(self.tokenizer) + idx] = '##' + chr(token)
-        self.tokenizer.added_tokens_decoder.update(suffix_vocab)
-        self._samples, self._global_index = self._read_data_files(data_path)
-
-    def _get_data_files(self, data_path):
-        # Get all prefix of .npy/.npz files in the current and next-level directories.
-        files = [
-            os.path.join(data_path, f) for f in os.listdir(data_path)
-            if (os.path.isfile(os.path.join(data_path, f)) and "_idx.npz" in
-                str(f))
-        ]
-        files = [x.replace("_idx.npz", "") for x in files]
-        return files
-
-    def _read_data_files(self, data_path):
-        data_files = self._get_data_files(data_path)
-        samples = []
-        indexes = []
-        for file_id, file_name in enumerate(data_files):
-
-            for suffix in ["_ids.npy", "_idx.npz"]:
-                if not os.path.isfile(file_name + suffix):
-                    raise ValueError("File Not found, %s" %
-                                     (file_name + suffix))
-
-            token_ids = np.load(
-                file_name + "_ids.npy", mmap_mode="r", allow_pickle=True)
-            samples.append(token_ids)
-
-            split_ids = np.load(file_name + "_idx.npz")
-            end_ids = np.cumsum(split_ids["lens"], dtype=np.int64)
-            file_ids = np.full(end_ids.shape, file_id)
-            split_ids = np.stack([file_ids, end_ids], axis=-1)
-            indexes.extend(split_ids)
-        indexes = np.stack(indexes, axis=0)
-        return samples, indexes
-
-    def __len__(self):
-        return len(self._global_index)
-
-    def __getitem__(self, index):
-        file_id, end_id = self._global_index[index]
-        start_id = 0
-        if index > 0:
-            pre_file_id, pre_end_id = self._global_index[index - 1]
-            if pre_file_id == file_id:
-                start_id = pre_end_id
-        word_token_ids = self._samples[file_id][start_id:end_id]
-        token_ids = []
-        is_suffix = np.zeros(word_token_ids.shape)
-        for idx, token_id in enumerate(word_token_ids):
-            token = self.tokenizer.convert_ids_to_tokens(int(token_id))
-            if '##' in token:
-                token_id = self.tokenizer.convert_tokens_to_ids(token[-1])
-                is_suffix[idx] = 1
-            token_ids.append(token_id)
-
-        return token_ids, is_suffix.astype(np.int64)
-
-
-class DataCollatorForErnieHealth(object):
-    def __init__(self, tokenizer, mlm_prob, max_seq_length):
-        self.tokenizer = tokenizer
-        self.mlm_prob = mlm_prob
-        self.max_seq_len = max_seq_length
-        self._ids = {
-            'cls':
-            self.tokenizer.convert_tokens_to_ids(self.tokenizer.cls_token),
-            'sep':
-            self.tokenizer.convert_tokens_to_ids(self.tokenizer.sep_token),
-            'pad':
-            self.tokenizer.convert_tokens_to_ids(self.tokenizer.pad_token),
-            'mask':
-            self.tokenizer.convert_tokens_to_ids(self.tokenizer.mask_token)
-        }
-
-    def __call__(self, data):
-        masked_input_ids_a, input_ids_a, labels_a = self.mask_tokens(data)
-        masked_input_ids_b, input_ids_b, labels_b = self.mask_tokens(data)
-        masked_input_ids = paddle.concat(
-            [masked_input_ids_a, masked_input_ids_b], axis=0).astype('int64')
-        input_ids = paddle.concat([input_ids_a, input_ids_b], axis=0)
-        labels = paddle.concat([labels_a, labels_b], axis=0)
-        return masked_input_ids, input_ids, labels
-
-    def mask_tokens(self, batch_data):
-
-        token_ids = [x[0] for x in batch_data]
-        is_suffix = [x[1] for x in batch_data]
-
-        # Create probability matrix where the probability of real tokens is
-        # self.mlm_prob, while that of others is zero.
-        data = self.add_special_tokens_and_set_maskprob(token_ids, is_suffix)
-        token_ids, is_suffix, prob_matrix = data
-        token_ids = paddle.to_tensor(
-            token_ids, dtype="int64", stop_gradient=True)
-        masked_token_ids = token_ids.clone()
-        labels = token_ids.clone()
-
-        # Create masks for words, where '百' must be masked if '度' is masked
-        # for the word '百度'.
-        prob_matrix = prob_matrix * (1 - is_suffix)
-        word_mask_index = np.random.binomial(1, prob_matrix).astype("float")
-        is_suffix_mask = (is_suffix == 1)
-        word_mask_index_tmp = word_mask_index
-        while word_mask_index_tmp.sum() > 0:
-            word_mask_index_tmp = np.concatenate(
-                [
-                    np.zeros((word_mask_index.shape[0], 1)),
-                    word_mask_index_tmp[:, :-1]
-                ],
-                axis=1)
-            word_mask_index_tmp = word_mask_index_tmp * is_suffix_mask
-            word_mask_index += word_mask_index_tmp
-        word_mask_index = word_mask_index.astype('bool')
-        labels[~word_mask_index] = -100
-
-        # 80% replaced with [MASK].
-        token_mask_index = paddle.bernoulli(paddle.full(
-            labels.shape, 0.8)).astype('bool').numpy() & word_mask_index
-        masked_token_ids[token_mask_index] = self._ids['mask']
-
-        # 10% replaced with random token ids.
-        token_random_index = paddle.to_tensor(
-            paddle.bernoulli(paddle.full(labels.shape, 0.5)).astype("bool")
-            .numpy() & word_mask_index & ~token_mask_index)
-        random_tokens = paddle.randint(
-            low=0,
-            high=self.tokenizer.vocab_size,
-            shape=labels.shape,
-            dtype='int64')
-        masked_token_ids = paddle.where(token_random_index, random_tokens,
-                                        masked_token_ids)
-
-        return masked_token_ids, token_ids, labels
-
-    def add_special_tokens_and_set_maskprob(self, token_ids, is_suffix):
-        batch_size = len(token_ids)
-        batch_token_ids = np.full((batch_size, self.max_seq_len),
-                                  self._ids['pad'])
-        batch_token_ids[:, 0] = self._ids['cls']
-        batch_is_suffix = np.full_like(batch_token_ids, -1)
-        prob_matrix = np.zeros_like(batch_token_ids, dtype='float32')
-
-        for idx in range(batch_size):
-            if len(token_ids[idx]) > self.max_seq_len - 2:
-                token_ids[idx] = token_ids[idx][:self.max_seq_len - 2]
-                is_suffix[idx] = is_suffix[idx][:self.max_seq_len - 2]
-            seq_len = len(token_ids[idx])
-            batch_token_ids[idx, seq_len + 1] = self._ids['sep']
-            batch_token_ids[idx, 1:seq_len + 1] = token_ids[idx]
-            batch_is_suffix[idx, 1:seq_len + 1] = is_suffix[idx]
-            prob_matrix[idx, 1:seq_len + 1] = self.mlm_prob
-
-        return batch_token_ids, batch_is_suffix, prob_matrix
-
-
-def create_dataloader(dataset,
-                      mode='train',
-                      batch_size=1,
-                      use_gpu=True,
-                      data_collator=None):
-    """
-    Creats dataloader.
-    Args:
-        dataset(obj:`paddle.io.Dataset`):
-            Dataset instance.
-        mode(obj:`str`, optional, defaults to obj:`train`):
-            If mode is 'train', it will shuffle the dataset randomly.
-        batch_size(obj:`int`, optional, defaults to 1): 
-            The sample number of a mini-batch.
-        use_gpu(obj:`bool`, optional, defaults to obj:`True`):
-            Whether to use gpu to run.
-    Returns:
-        dataloader(obj:`paddle.io.DataLoader`): The dataloader which generates batches.
-    """
-
-    if mode == 'train' and use_gpu:
-        sampler = paddle.io.DistributedBatchSampler(
-            dataset=dataset, batch_size=batch_size, shuffle=True)
-        dataloader = paddle.io.DataLoader(
-            dataset,
-            batch_sampler=sampler,
-            return_list=True,
-            collate_fn=data_collator,
-            num_workers=0)
-    else:
-        shuffle = True if mode == 'train' else False
-        sampler = paddle.io.BatchSampler(
-            dataset=dataset, batch_size=batch_size, shuffle=shuffle)
-        dataloader = paddle.io.DataLoader(
-            dataset,
-            batch_sampler=sampler,
-            return_list=True,
-            collate_fn=data_collator,
-            num_workers=0)
-
-    return dataloader
-
-
 def do_train(args):
     paddle.enable_static() if not args.eager_run else None
     paddle.set_device(args.device)
@@ -394,8 +171,7 @@ def do_train(args):
     set_seed(args.seed)
     worker_init = WorkerInitObj(args.seed + paddle.distributed.get_rank())
 
-    args.model_type = args.model_type.lower()
-    model_class, tokenizer_class = MODEL_CLASSES[args.model_type]
+    model_class, tokenizer_class = MODEL_CLASSES['ernie-health']
 
     # Loads or initialize a model.
     pretrained_models = list(tokenizer_class.pretrained_init_configuration.keys(
@@ -456,25 +232,27 @@ def do_train(args):
 
     # Loads dataset.
     tic_load_data = time.time()
-    print("start load data : %s" %
-          (time.strftime("%Y-%m-%d %H:%M:%S", time.localtime())))
+    logger.info("start load data : %s" %
+                (time.strftime("%Y-%m-%d %H:%M:%S", time.localtime())))
 
     train_dataset = MedicalCorpus(data_path=args.input_dir, tokenizer=tokenizer)
-    print("load data done, total : %s s" % (time.time() - tic_load_data))
+    logger.info("load data done, total : %s s" % (time.time() - tic_load_data))
 
     # Reads data and generates mini-batches.
     data_collator = DataCollatorForErnieHealth(
-        tokenizer=tokenizer, max_seq_length=args.max_seq_length, mlm_prob=0.15)
+        tokenizer=tokenizer,
+        max_seq_length=args.max_seq_length,
+        mlm_prob=args.mlm_prob)
 
     train_data_loader = create_dataloader(
         train_dataset,
-        batch_size=args.train_batch_size,
+        batch_size=args.batch_size,
         mode='train',
         use_gpu=True if args.device in "gpu" else False,
         data_collator=data_collator)
 
     num_training_steps = args.max_steps if args.max_steps > 0 else (
-        len(train_data_loader) * args.num_train_epochs)
+        len(train_data_loader) * args.num_epochs)
 
     lr_scheduler = LinearDecayWithWarmup(args.learning_rate, num_training_steps,
                                          args.warmup_steps)
@@ -497,8 +275,8 @@ def do_train(args):
     if args.use_amp:
         scaler = paddle.amp.GradScaler(init_loss_scaling=1024)
 
-    print("start train : %s" %
-          (time.strftime("%Y-%m-%d %H:%M:%S", time.localtime())))
+    logger.info("start train : %s" %
+                (time.strftime("%Y-%m-%d %H:%M:%S", time.localtime())))
     trained_global_step = global_step = 0
     t_loss = defaultdict(lambda: paddle.to_tensor([0.0]))
     log_loss = defaultdict(lambda: paddle.to_tensor([0.0]))
@@ -512,19 +290,19 @@ def do_train(args):
                 os.path.join(args.model_name_or_path, "model_state.pdopt")))
         trained_global_step = global_step = config_dict["global_step"]
         if trained_global_step < num_training_steps:
-            print(
+            logger.info(
                 "[ start train from checkpoint ] we have already trained %s steps, seeking next step : %s"
                 % (trained_global_step, trained_global_step + 1))
         else:
-            print(
+            logger.info(
                 "[ start train from checkpoint ] we have already trained %s steps, but total training steps is %s, please check configuration !"
                 % (trained_global_step, num_training_steps))
             exit(0)
 
     if paddle.distributed.get_rank() == 0:
-        writer = SummaryWriter(os.path.join(args.output_dir, 'log'))
+        writer = LogWriter(os.path.join(args.output_dir, 'loss_log'))
 
-    for epoch in range(args.num_train_epochs):
+    for epoch in range(args.num_epochs):
         for step, batch in enumerate(train_data_loader):
             if trained_global_step > 0:
                 trained_global_step -= 1
@@ -591,7 +369,7 @@ def do_train(args):
                                  tmp_loss['csp'],
                                  optimizer.get_lr(),
                                  (time.time() - tic_train) / args.logging_steps)
-                        print(log_str)
+                        logger.info(log_str)
                         log_list.append(log_str)
                         writer.add_scalar('generator_loss', tmp_loss['gen'],
                                           global_step)
@@ -618,7 +396,7 @@ def do_train(args):
                              local_loss['csp'],
                              optimizer.get_lr(),
                              (time.time() - tic_train) / args.logging_steps)
-                    print(log_str)
+                    logger.info(log_str)
                     log_list.append(log_str)
                     writer.add_scalars('loss', {
                         'generator_loss': local_loss['gen'],
@@ -672,6 +450,8 @@ def do_train(args):
                                 if len(log.strip()) > 0:
                                     f.write(log.strip() + '\n')
             if global_step >= num_training_steps:
+                if paddle.distributed.get_rank() == 0:
+                    writer.close()
                 return
